requirements.txt
pandas
numpy
numexpr
python-dotenv
requests
sqlalchemy
pyodbc

networkx